+++ awc/src/error.rs
@@ -11,6 +11,8 @@ use serde_json::error::Error as JsonError;
 
 pub use crate::client::{ConnectError, FreezeRequestError, InvalidUrl, SendRequestError};
 
+// TODO: address display, error, and from impls
+
 /// Websocket client error
 #[derive(Debug, Display, From)]
 pub enum WsClientError {